[12.0][REF] website_sale_customer_type: externalize signup externalization done in `website_sale_customer_type_signup`
