+++ website_sale_customer_type/views/website_sale_templates.xml
@@ -36,38 +36,6 @@
         </div>
     </template>
 
-    <!-- TODO: fix when module is split into `website_sale_customer_type_signup` -->
-    <!--
-    <template id="customer_type_address" inherit_id="website_sale.address">
-        <xpath expr="//input[@name='vat']" position="attributes">
-            <t t-if="customer_type and customer_type.website_signup_vat_required">
-                <attribute name="t-att-required">'required' if (customer_type and customer_type.website_signup_vat_required) else None</attribute>
-            </t>
-        </xpath>
-        <xpath expr="//form[@name='/shop/address']//t[@t-if=&quot;mode == (&apos;new&apos;, &apos;billing&apos;)&quot;]" position="after">
-            <t t-if="mode == ('edit', 'billing')">
-                <t t-if="customer_type and customer_type.website_signup_vat_required">
-                    <div t-attf-class="form-group #{error.get('company_name') and 'has-error' or ''} col-md-6">
-                        <label class="control-label font-weight-normal label-optional" for="company_name">Company Name</label>
-                        <input type="text"
-                               name="company_name"
-                               class="form-control"
-                               t-att-value="'company_name' in checkout and checkout['company_name']" />
-                    </div>
-                    <div t-attf-class="form-group #{error.get('vat') and 'has-error' or ''} col-md-6 div_vat">
-                        <label class="control-label" for="vat">TIN / VAT </label>
-                        <input type="text"
-                               name="vat"
-                               class="form-control"
-                               t-att-value="'vat' in checkout and checkout['vat']"
-                               t-att-required="'required' if (customer_type and customer_type.website_signup_vat_required) else None" />
-                    </div>
-                </t>
-            </t>
-        </xpath>
-    </template>
-    -->
-
     <template id="customer_type_selector_popover_bottom" name="Customer Type Selector Popover Bottom">
         <!-- Edit this template to add content at the bottom of the customer selector -->
     </template>
