+++ website_sale_customer_type_signup/README.rst
@@ -1,0 +1,66 @@
+=================================
+Website Sale Customer Type Signup
+=================================
+
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-coopiteasy%2Fcie--e--commerce-lightgray.png?logo=github
+    :target: https://github.com/coopiteasy/cie-e-commerce/tree/12.0/website_sale_customer_type_signup
+    :alt: coopiteasy/cie-e-commerce
+
+|badge1| |badge2| |badge3| 
+
+This module:
+
+* Specify whether a Customer Type is allowed to create an account on e-commerce
+* Specify whether a Customer Type must specify the TIN/VAT number on signup
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Known issues / Roadmap
+======================
+
+- module test
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/coopiteasy/cie-e-commerce/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/coopiteasy/cie-e-commerce/issues/new?body=module:%20website_sale_customer_type_signup%0Aversion:%2012.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
+
+Credits
+=======
+
+Authors
+~~~~~~~
+
+* Coop IT Easy SCRLfs <https://coopiteasy.be>
+
+Contributors
+~~~~~~~~~~~~
+
+* Rémy Taymans <remy@coopiteasy.be>
+* Vincent Van Rossem <vincent@coopiteasy.be>
+
+Maintainers
+~~~~~~~~~~~
+
+This module is part of the `coopiteasy/cie-e-commerce <https://github.com/coopiteasy/cie-e-commerce/tree/12.0/website_sale_customer_type_signup>`_ project on GitHub.
+
+You are welcome to contribute.
