+++ website_sale_customer_type_signup/__manifest__.py
@@ -2,24 +2,24 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 {
-    'name': 'Website Sale Customer Type',
+    'name': 'Website Sale Customer Type Signup',
     'summary': """
-        Let customer choose his type when accessing the e-commerce""",
+        Restrict Customer Type Signup on E-commerce""",
     'version': '12.0.1.0.0',
     'license': 'AGPL-3',
     'author': 'Coop IT Easy SCRLfs <https://coopiteasy.be>',
     'website': 'https://github.com/coopiteasy/cie-e-commerce',
     'depends': [
-        "contacts",
-        "website_sale",
+        "auth_signup",
+        "website_sale_customer_type",
     ],
     'data': [
-        "views/product_product.xml",
-        "views/res_partner.xml",
+        "views/auth_signup_login_templates.xml",
+        "views/portal_templates.xml",
         "views/res_partner_customer_type.xml",
         "views/website_sale_templates.xml",
-        "security/ir.model.access.csv",
     ],
     'demo': [
     ],
+    'installable': False,  # not tested
 }
