+++ website_sale_customer_type_signup/controllers/__init__.py
@@ -1,0 +1,1 @@
+from . import main
