+++ website_sale_customer_type_signup/models/__init__.py
@@ -1,0 +1,2 @@
+from . import res_partner_customer_type
+from . import res_users
