+++ website_sale_customer_type_signup/models/res_users.py
