+++ website_sale_customer_type_signup/readme/CONTRIBUTORS.rst
@@ -1,0 +1,2 @@
+* Rémy Taymans <remy@coopiteasy.be>
+* Vincent Van Rossem <vincent@coopiteasy.be>
