+++ website_sale_customer_type_signup/readme/DESCRIPTION.rst
@@ -1,0 +1,4 @@
+This module:
+
+* Specify whether a Customer Type is allowed to create an account on e-commerce
+* Specify whether a Customer Type must specify the TIN/VAT number on signup
