+++ website_sale_customer_type_signup/views/auth_signup_login_templates.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<!-- Copyright 2018-2020 Coop IT Easy SCRLfs
+<!-- Copyright 2021 Coop IT Easy SCRLfs
      License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl). -->
 <odoo>
     <template id="customer_type_fields" inherit_id="auth_signup.fields">
